requirements.txt
environs==9.5.0
requests==2.31.0
python-telegram-bot==13.14
dataclasses==0.6
